Some cleanup and typo fixes

+++ 9-post-setup.sh
@@ -5,49 +5,6 @@ echo "## FINAL SETUPS ##"
 echo "##################"
 echo
 
-DOTDIR=$HOME/RinnOS/dotfiles
-# DOTFILES=(
-#     # --- Home ---
-#     '.zshrc'
-#     '.gitconfig'
-#     '.emacs.d'
-#     '.doom.d'
-#     'wallpapers'
-#     'scripts'
-
-#     # --- .config ---
-#     '.config/i3'
-#     '.config/pulsemeeter'
-#     '.config/flameshot'
-# )
-
-LOCALDOTDIR=/mnt/Vault/Dotfiles/Linux/Arch
-# LOCALDOTFILES=(
-#     # --- Home ---
-#     '.ssh'
-#     '.waterfox'
-#     '.vscode'
-
-#     # --- .config ---
-#     '.config/BetterDiscord'
-#     '.config/Code'
-#     '.config/discord'
-#     '.config/discordcanary'
-#     '.config/FreeTube'
-#     '.config/Logseq'
-#     '.config/lutris'
-#     '.config/Rambox'
-#     '.config/rambox'
-#     '.config/wtwitch'
-
-#     # --- .local ---
-#     '.local/share/lutris'
-#     '.local/share/Steam'
-#     '.local/share/vulkan'
-#     '.local/share/PolyMC'
-#     '.local/share/7DaysToDie'
-# )
-
 echo "-----------------------"
 echo "-- Starting Services --"
 echo "-----------------------"
@@ -79,6 +36,10 @@ I=0
 while [ $I -le $(yq '.git | length' dotfiles.yaml) ]; do
     dotfile=$(yq .git[$I] dotfiles.yaml)
 
+    if [ -z "$dotfile" ]; then
+        break
+    fi
+
     rm -rf "${HOME}/$dotfile"
     echo
     echo "Linking ${DOTDIR}/${dotfile} to ${HOME}/${dotfile}"
@@ -86,16 +47,13 @@ while [ $I -le $(yq '.git | length' dotfiles.yaml) ]; do
     ((I++))
 done
 
-# for dotfile in "${DOTFILES[@]}"; do
-#     rm -rf "${HOME}/$dotfile"
-#     echo
-#     echo "Linking ${DOTDIR}/${dotfile} to ${HOME}/${dotfile}"
-#     cp -faTs "${DOTDIR}/${dotfile}" "${HOME}/${dotfile}"
-# done
-
 I=0
 while [ $I -le $(yq '.local | length' dotfiles.yaml) ]; do
     localdotfile=$(yq .local[$I] dotfiles.yaml)
+
+    if [ -z "$localdotfile" ]; then
+        break
+    fi
 
     rm -rf "${HOME}/${localdotfile}"
     echo
@@ -103,13 +61,6 @@ while [ $I -le $(yq '.local | length' dotfiles.yaml) ]; do
     cp -faTs "${LOCALDOTDIR}/${localdotfile}" "${HOME}/${localdotfile}"
     ((I++))
 done
-
-# for localdotfile in "${LOCALDOTFILES[@]}"; do
-#     rm -rf "${HOME}/${localdotfile}"
-#     echo
-#     echo "Linking ${LOCALDOTDIR}/${localdotfile} to ${HOME}/${localdotfile}"
-#     cp -faTs "${LOCALDOTDIR}/${localdotfile}" "${HOME}/${localdotfile}"
-# done
 
 echo "-------------------"
 echo "-- Virtulazation --"
